added postgres db

+++ services/booking_service/que/handlers.go
@@ -36,7 +36,7 @@ func GenerateSeatsHandler(
 				seats = append(seats, domain.Seat{
 					EventID:    msg.EventID,
 					RowID:      r.Row,
-					SeatNumber: i,
+					SeatNumber: int32(i),
 					Status:     "AVAILABLE",
 				})
 			}
